Fixed a bug on the release administration page where the actual release date was referred to as planned release date.
Made the issue ID on the issue view a bookmarkable link so that people 
could easily copy/paste the link for sharing or use it for bookmarking.  
The view's URL loses its nice bookmarkable form after an attachment has 
been made and in some other instances.

Added a search by project search item.